# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "fsGaze API",
    "description": "API for analyzing KerML documents and managing safety analysis data.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/safety/risk-ratings/{uuid}": {
        "get": {
            "tags": ["Risk Ratings"],
            "summary": "Get a risk rating by UUID",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "put": {
            "tags": ["Risk Ratings"],
            "summary": "Update a risk rating",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateRiskRating"
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Update successful" },
                "400": { "description": "Bad request" }
            }
        },
        "delete": {
            "tags": ["Risk Ratings"],
            "summary": "Delete a risk rating",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Deletion successful" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
    "/api/safety/safety-requirements": {
        "get": {
            "tags": ["Safety Requirements"],
            "summary": "Get all safety requirements",
            "parameters": [
              {
                "in": "query",
                "name": "reqASIL",
                "schema": { "type": "string" },
                "description": "Filter by ASIL level"
              },
              {
                "in": "query",
                "name": "name",
                "schema": { "type": "string" },
                "description": "Filter by name (partial match)"
              },
              {
                "in": "query",
                "name": "reqID",
                "schema": { "type": "string" },
                "description": "Filter by requirement ID (partial match)"
              }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" }
            }
        }
    },
    "/api/safety/safety-requirements/{uuid}": {
        "get": {
            "tags": ["Safety Requirements"],
            "summary": "Get a safety requirement by UUID",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "put": {
            "tags": ["Safety Requirements"],
            "summary": "Update a safety requirement",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateSafetyRequirement"
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Update successful" },
                "400": { "description": "Bad request" }
            }
        },
        "delete": {
            "tags": ["Safety Requirements"],
            "summary": "Delete a safety requirement",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Deletion successful" },
                "400": { "description": "Bad request" }
            }
        }
    },
    "/api/safety/risk-ratings/{uuid}/tasks": {
        "get": {
            "tags": ["Safety Tasks"],
            "summary": "Get safety tasks for a risk rating",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "post": {
            "tags": ["Safety Tasks"],
            "summary": "Create a safety task for a risk rating",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateSafetyTask"
                        }
                    }
                }
            },
            "responses": {
                "201": { "description": "Safety task created" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
    "/api/safety/safety-tasks/{uuid}": {
        "get": {
            "tags": ["Safety Tasks"],
            "summary": "Get a safety task by UUID",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "put": {
            "tags": ["Safety Tasks"],
            "summary": "Update a safety task",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateSafetyTask"
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Update successful" },
                "400": { "description": "Bad request" }
            }
        },
        "delete": {
            "tags": ["Safety Tasks"],
            "summary": "Delete a safety task",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Deletion successful" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
    "/api/safety/safety-tasks/{uuid}/notes": {
        "get": {
            "tags": ["Safety Notes"],
            "summary": "Get notes for a safety task",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "post": {
            "tags": ["Safety Notes"],
            "summary": "Create a note for a safety task",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateSafetyNote"
                        }
                    }
                }
            },
            "responses": {
                "201": { "description": "Note created" },
                "400": { "description": "Bad request" }
            }
        }
    },
    "/api/safety/safety-notes/{uuid}": {
        "get": {
            "tags": ["Safety Notes"],
            "summary": "Get a safety note by UUID",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "put": {
            "tags": ["Safety Notes"],
            "summary": "Update a safety note",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateSafetyNote"
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Update successful" },
                "400": { "description": "Bad request" }
            }
        },
        "delete": {
            "tags": ["Safety Notes"],
            "summary": "Delete a safety note",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Deletion successful" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRiskRating": {
          "type": "object",
          "properties": {
              "severity": { 
                "type": "integer",
                "minimum": 1,
                "maximum": 10,
                "description": "Severity rating (1-10)",
                "example": 7
              },
              "occurrence": { 
                "type": "integer",
                "minimum": 1,
                "maximum": 10,
                "description": "Occurrence rating (1-10)",
                "example": 3
              },
              "detection": { 
                "type": "integer",
                "minimum": 1,
                "maximum": 10,
                "description": "Detection rating (1-10)",
                "example": 5
              },
              "ratingComment": { 
                "type": "string",
                "description": "Optional comment about the risk rating",
                "example": "Based on historical data and expert judgment"
              }
          },
          "required": ["severity", "occurrence", "detection"]
      },
...
      "UpdateSafetyRequirement": {
          "type": "object",
          "properties": {
            "name": { "type": "string" },
            "reqID": { "type": "string" },
            "reqText": { "type": "string" },
            "reqASIL": { 
              "type": "string",
              "enum": ["QM", "A", "B", "C", "D"]
            },
            "reqLinkedTo": { "type": "string" }
          }
      },
      "CreateSafetyTask": {
          "type": "object",
          "properties": {
              "name": { 
                "type": "string",
                "description": "Name of the safety task",
                "example": "Implement watchdog timer"
              },
              "description": { 
                "type": "string",
                "description": "Detailed description of the task",
                "example": "Implement a hardware watchdog timer to detect software failures"
              },
              "status": { 
                "type": "string",
                "enum": ["open", "started", "in-review", "finished"],
                "description": "Current status of the task",
                "example": "open"
              },
              "responsible": { 
                "type": "string",
                "description": "Person or team responsible for the task",
                "example": "Safety Team"
              },
              "reference": { 
                "type": "string",
                "description": "Reference or ticket number",
                "example": "JIRA-1234"
              },
              "taskType": { 
                "type": "string",
                "enum": ["runtime measures", "dev-time measures", "other"],
                "description": "Type of safety task",
                "example": "runtime measures"
              }
          },
          "required": ["name", "description", "status", "responsible", "reference", "taskType"]
      },
...
      "CreateSafetyNote": {
          "type": "object",
          "properties": {
              "note": { 
                "type": "string",
                "description": "Content of the safety note",
                "example": "This failure mode requires special attention during testing phase"
              }
          },
          "required": ["note"]
      }
    }
  }
}